+++ src/tool/MethyProfile/profile_node.h
@@ -13,8 +13,8 @@ struct BasicEntry
     char str_type[ENTRY_TYPE_BUFFER_LENGTH] = {0};  /* Type in an entry. */
     int chr = 0;                /* The chromosome where the gene is located. */
     bool chain = -1;            /* Positive chain (true) or negative chain (false). */
-    unsigned long Start= 0;
-    unsigned long End= 0;       /* Location on chromosome. */
+    unsigned long long Start= 0;
+    unsigned long long End= 0;       /* Location on chromosome. */
     double methy_ratio = 0.0f;  /* Methylation ratio of this gene. */
 };
 
@@ -34,8 +34,8 @@ struct ProfileNode : BasicEntry
     char ID[GENE_ID_BUFFER_LENGTH] = {0};           /* ID of the gene. */
     bool single_tag = false;                        /* TRUE when this gene will be analysed singly. */
 #ifdef CG_NUMBER
-    unsigned long NumCG = 0;            /* CG methylation number. */
-    unsigned long NumCG_promoter = 0;   /* CG methylation number in promoter. */
+    unsigned long long NumCG = 0;            /* CG methylation number. */
+    unsigned long long NumCG_promoter = 0;   /* CG methylation number in promoter. */
 #endif// !CG_NUMBER
     double methy_ratio_promoter = 0.0f;
         /* Methylation ratio of the promoter of this gene. */
